requirements.txt
click
coloredlogs
verboselogs
python-decouple
